requirements.txt
Django>=4.0
djangorestframework>=3.13
djangorestframework-simplejwt>=5.2
drf-spectacular>=0.26
django-filter
django-cors-headers>=3.13
mysqlclient>=2.1
PyJWT>=2.4
Pillow>=9.0
python-dotenv>=1.0
